feat: created a shelftoshef method to change books within racks
This method will take two event as a parameter id and shelfvalue
After that will pass that id to check the value and replace with the target value.
Then will render the books array based on the shelf value and rack.

Now the app is working like:
we can search books
we can add search books to any desired shelf
we can exchange the books within different shelfs.