feat: both shelf relected same if change makes in search page
If the book shelf is change on search page,the change will reflect on main page by checkin the id and shelf.
If the id is fount in main page will change the shel and update the books state else will add the new book to the main page and update the state.