fix: error and warning checked
Applied and fixed all different keys and ids in each list of iteration by passind index as parametre.
Change selected attribute to default value in select options.

style: Added zoom

This gives a proper view of books in page  to the viewer

feat: number of books will display while searching

Added length of state so that it will display the total count of search reluts books.